+++ js/components/render-load-photo.js
@@ -6,7 +6,7 @@ import {
   removeEffectClass,
   resetScale,
 } from './render-scale.js';
-import { closeModal, setModal } from './modal-handler.js';
+import { closeModal, setModal } from '../store/modal-handler.js';
 
 const body = document.body;
 const imgUpload = body.querySelector('.img-upload');
